+++ src/hybrid-appliance/azext_hybrid_appliance/custom.py
@@ -89,9 +89,18 @@ def validate_hybrid_appliance(cmd, resource_group_name, name, validate_connected
             cmd_show_arc= ['az', 'connectedk8s', 'show', '-n', name, '-g', resource_group_name, '-o', 'none']
             process = subprocess.Popen(cmd_show_arc, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
             if process.wait() == 0:
-                telemetry.set_exception(exception="An appliance with same name already exists in the resource group", fault_type=consts.Resource_Already_Exists_Fault_Type, summary="Appliance resource with same name already exists")
-                logger.warning("The appliance name and resource group provided in the script already correspond to an existing connected cluster resource in Azure. Go to the portal and generate the script again with a different appliance name.")
-                all_validations_passed = False
+                matches_current_cluster=False
+                if utils.check_microk8s():
+                    try:
+                        cm = utils.get_azure_clusterconfig_cm()
+                        utils.validate_cluster_resource_group_and_name(cm, resource_group_name, name)
+                        matches_current_cluster=True
+                    except:
+                        pass
+                if not matches_current_cluster:
+                    telemetry.set_exception(exception="An appliance with same name already exists in the resource group", fault_type=consts.Resource_Already_Exists_Fault_Type, summary="Appliance resource with same name already exists")
+                    logger.warning("The appliance name and resource group provided in the script already correspond to an existing connected cluster resource in Azure. Go to the portal and generate the script again with a different appliance name.")
+                    all_validations_passed = False
             stdout = process.stdout.read().decode()
             stderr = process.stderr.read().decode()
             if "ResourceGroupNotFound" in stdout or "ResourceGroupNotFound" in stderr:
@@ -111,8 +120,17 @@ def validate_hybrid_appliance(cmd, resource_group_name, name, validate_connected
     else:
         print("All pre-requisite validations have passed successfully")
 
-def create_hybrid_appliance(cmd, resource_group_name, name, correlation_id=None, https_proxy="", http_proxy="", no_proxy="", proxy_cert="", location=None, tags=None):
+def create_hybrid_appliance(cmd, resource_group_name, name, correlation_id=None, https_proxy="", http_proxy="", no_proxy="", proxy_cert="", location=None, tags=None, custom_locations_oid=""):
     if utils.check_microk8s(throw_on_subprocess_exception=False):
+        matches_current_cluster=False
+        try:
+            cm = utils.get_azure_clusterconfig_cm()
+            utils.validate_cluster_resource_group_and_name(cm, resource_group_name, name)
+            matches_current_cluster=True
+        except:
+            pass
+        if matches_current_cluster:
+            return
         raise ValidationError("A MicroK8s cluster is already running on this server. Run the script in “delete” mode to clean-up any existing components and execute the script again. Learn more: <URL>")
 
     kubeconfig_path = utils.get_kubeconfig_path()
@@ -180,6 +198,8 @@ def create_hybrid_appliance(cmd, resource_group_name, name, correlation_id=None,
         cmd_onboard_arc.extend(["--tags"])
         for tag in tags:
             cmd_onboard_arc.extend(["{}={}".format(tag, tags[tag])])
+    if custom_locations_oid:
+        cmd_onboard_arc.extend(["--custom-locations-oid", custom_locations_oid])
 
     onboarding_result = get_default_cli().invoke(cmd_onboard_arc)
     if onboarding_result != 0: