+++ src/hybrid-appliance/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.1.4
+++++++
+* Updated prechecks and onboarding to not fail if appliance already exists with same name and rg.
+* Added CL OID as a parameter for hybrid appliance create to allow SPN onboarding
+
 0.1.3
 ++++++
 Added check for cpu cores